refactor: ViewHolder 분리 및 패키지 구조 변경
