+++ app/src/main/java/woowacourse/movie/movie/MovieFragment.kt
@@ -11,6 +11,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import woowacourse.movie.R
 import woowacourse.movie.databinding.FragmentMovieBinding
 import woowacourse.movie.movie.adapter.MovieAdapter
+import woowacourse.movie.theater.TheaterFragment
 import woowacourse.movie.ui.model.MovieUiModel
 import woowacourse.movie.ui.model.TheaterUiModel
