+++ app/src/main/java/woowacourse/movie/movie/adapter/MovieViewHolder.kt
@@ -1,0 +1,13 @@
+package woowacourse.movie.movie.adapter
+
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.movie.databinding.MovieListItemBinding
+import woowacourse.movie.ui.model.MovieUiModel
+
+class MovieViewHolder(val binding: MovieListItemBinding, val onReserveClick: ReserveClickListener) :
+    RecyclerView.ViewHolder(binding.root) {
+    fun bind(movie: MovieUiModel) {
+        binding.movie = movie
+        binding.clickListener = onReserveClick
+    }
+}
