+++ app/src/main/java/woowacourse/movie/theater/TheaterAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.movie.adapter
+package woowacourse.movie.theater
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -9,14 +9,14 @@ import woowacourse.movie.ui.model.TheaterUiModel
 class TheaterAdapter(
     private val theaters: List<TheaterUiModel>,
     private val onSelectClick: SelectClickListener,
-) : RecyclerView.Adapter<TheaterAdapter.TheaterViewHolder>() {
+) : RecyclerView.Adapter<TheaterViewHolder>() {
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int,
     ): TheaterViewHolder {
         val inflater = LayoutInflater.from(parent.context)
         val binding = TheaterItemBinding.inflate(inflater, parent, false)
-        return TheaterViewHolder(binding)
+        return TheaterViewHolder(binding, onSelectClick)
     }
 
     override fun getItemCount(): Int = theaters.size
@@ -28,13 +28,5 @@ class TheaterAdapter(
         val theater = theaters[position]
 
         holder.bind(theater)
-    }
-
-    inner class TheaterViewHolder(val binding: TheaterItemBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(theater: TheaterUiModel) {
-            binding.theater = theater
-            binding.clickListener = onSelectClick
-        }
     }
 }
