+++ app/src/main/java/woowacourse/movie/theater/TheaterFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.movie
+package woowacourse.movie.theater
 
 import android.content.Intent
 import android.os.Build
@@ -14,10 +14,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import woowacourse.movie.R
 import woowacourse.movie.booking.detail.BookingDetailActivity
-import woowacourse.movie.booking.detail.BookingDetailActivity.Companion.KEY_MOVIE_DATA
-import woowacourse.movie.booking.detail.BookingDetailActivity.Companion.KEY_THEATER_DATA
 import woowacourse.movie.databinding.FragmentTheaterBinding
-import woowacourse.movie.movie.adapter.TheaterAdapter
 import woowacourse.movie.ui.model.MovieUiModel
 import woowacourse.movie.ui.model.TheaterUiModel
 
@@ -50,8 +47,8 @@ class TheaterFragment : BottomSheetDialogFragment() {
                     setReorderingAllowed(true)
                     val intent =
                         Intent(activity, BookingDetailActivity::class.java).apply {
-                            putExtra(KEY_THEATER_DATA, theater)
-                            putExtra(KEY_MOVIE_DATA, initMovie())
+                            putExtra(BookingDetailActivity.Companion.KEY_THEATER_DATA, theater)
+                            putExtra(BookingDetailActivity.Companion.KEY_MOVIE_DATA, initMovie())
                         }
                     startActivity(intent)
                     addToBackStack(null)
