+++ app/src/main/res/layout/fragment_theater.xml
@@ -5,7 +5,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".movie.TheaterFragment">
+        tools:context=".theater.TheaterFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:padding="8dp"
